[local-to-server] refactor : Repository를 Category와 Memo로 분리
