+++ core/data/src/main/java/com/hegunhee/newsimplememoapp/data/dataSource/LocalDataSource.kt
@@ -7,13 +7,7 @@ interface LocalDataSource {
 
     suspend fun insertMemo(memo : MemoEntity)
 
-    suspend fun insertCategory(categoryEntity : CategoryEntity)
-
     suspend fun getMemo(memoId : Int) : MemoEntity
-
-    suspend fun getAllCategoryByType(categoryCode : Int) : List<CategoryEntity>
-
-    suspend fun checkIsCategory(categoryCode: Int, text : String) : CategoryEntity?
 
     suspend fun getMemoListSortedByYearAndMonth(year : Int, month : Int) : List<MemoEntity>
 
@@ -24,6 +18,12 @@ interface LocalDataSource {
     suspend fun deleteAllMemo()
 
     suspend fun deleteMemo(id : Int)
+
+    suspend fun getAllCategoryByType(categoryCode : Int) : List<CategoryEntity>
+
+    suspend fun checkIsCategory(categoryCode: Int, text : String) : CategoryEntity?
+
+    suspend fun insertCategory(categoryEntity : CategoryEntity)
 
     suspend fun deleteCategory(text : String)
