+++ core/data/src/main/java/com/hegunhee/newsimplememoapp/data/mapper/EntityMapper.kt
@@ -4,7 +4,7 @@ import com.hegunhee.newsimplememoapp.data.entity.MemoEntity
 import com.hegunhee.newsimplememoapp.domain.model.MemoType
 
 fun MemoEntity.toMemo() : MemoType.Memo {
-    return MemoType.Memo(category, year, month, day, dayOfWeek, amPm, hour, minute, attr, price, asset, description, id)
+    return MemoType.Memo(id,category, year, month, day, dayOfWeek, amPm, hour, minute, attr, price, asset, description)
 }
 
 fun MemoType.Memo.toMemoEntity() : MemoEntity {
