+++ core/data/src/main/java/com/hegunhee/newsimplememoapp/data/repository/DefaultMemoRepository.kt
@@ -2,6 +2,7 @@ package com.hegunhee.newsimplememoapp.data.repository
 
 import com.hegunhee.newsimplememoapp.data.dataSource.LocalDataSource
 import com.hegunhee.newsimplememoapp.data.entity.CategoryEntity
+import com.hegunhee.newsimplememoapp.data.entity.MemoEntity
 import com.hegunhee.newsimplememoapp.data.mapper.*
 import com.hegunhee.newsimplememoapp.domain.model.CategoryType
 import com.hegunhee.newsimplememoapp.domain.model.MemoType
@@ -15,21 +16,22 @@ class DefaultMemoRepository @Inject constructor(
     private val localDataSource: LocalDataSource
 ) : MemoRepository {
 
-    override suspend fun insertMemo(memo: MemoType.Memo) {
-        localDataSource.insertMemo(memo.toMemoEntity())
-    }
-
-    override suspend fun insertCategory(categoryType: CategoryType, text: String) {
-        localDataSource.insertCategory(CategoryEntity(categoryType.code,text))
-    }
-
     override suspend fun getMemo(memoId: Int) : MemoType.Memo {
         return localDataSource.getMemo(memoId).toMemo()
     }
 
     override suspend fun getMemoTypeListSortedByYearAndMonth(year: Int, month: Int): List<MemoType> {
         val memoList = localDataSource.getMemoListSortedByYearAndMonth(year,month)
-        return memoList
+        return memoList.getMonthMemoList(year,month);
+    }
+
+    override suspend fun getMemoTypeListByAttr(attr: String, year: Int, month: Int): List<MemoType> {
+        val memoList = localDataSource.getMemoListSortedByAttrYearMonth(attr,year,month)
+        return memoList.getMonthMemoList(year,month);
+    }
+
+    private fun List<MemoEntity>.getMonthMemoList(year : Int,month : Int) : List<MemoType> {
+        return this
             .groupBy { it.day }
             .flatMap { (day, list) ->
                 val incomeSum = list.filter { it.category == "수입" }.sumOf { it.price }
@@ -37,6 +39,22 @@ class DefaultMemoRepository @Inject constructor(
                 val memoDate = MemoType.MemoDate(year,month,day,list.firstOrNull()?.dayOfWeek ?: "월", incomeSum,expenseSum)
                 listOf(memoDate) + list.map { it.toMemo() }
             }
+    }
+
+    override suspend fun insertMemo(memo: MemoType.Memo) {
+        localDataSource.insertMemo(memo.toMemoEntity())
+    }
+
+    override suspend fun updateMemo(memo: MemoType.Memo) {
+        localDataSource.updateMemo(memo.toMemoEntity())
+    }
+
+    override suspend fun deleteMemo(id : Int) {
+        localDataSource.deleteMemo(id)
+    }
+
+    override suspend fun deleteAllMemo() {
+        localDataSource.deleteAllMemo()
     }
 
     override suspend fun getStaticsData(year: Int, month: Int): List<StaticsData> {
@@ -51,42 +69,5 @@ class DefaultMemoRepository @Inject constructor(
             }
             StaticsData(list[0].category,percent.toInt(),attr,list.sumOf { it.price },year,month)
         }
-    }
-
-    override suspend fun getMemoTypeListByAttr(attr: String, year: Int, month: Int): List<MemoType> {
-        val memoList = localDataSource.getMemoListSortedByAttrYearMonth(attr,year,month)
-        return memoList
-            .groupBy { it.day }
-            .flatMap { (day, list) ->
-                val incomeSum = list.filter { it.category == "수입" }.sumOf { it.price }
-                val expenseSum = list.filter { it.category == "지출"}.sumOf { it.price }
-                val memoDate = MemoType.MemoDate(year,month,day,list.firstOrNull()?.dayOfWeek ?: "월", incomeSum,expenseSum)
-                listOf(memoDate) + list.map { it.toMemo() }
-            }
-    }
-
-    override suspend fun getAllCategoryByType(categoryType: CategoryType): List<String> {
-        return localDataSource.getAllCategoryByType(categoryType.code).map { it.text }
-    }
-
-    override suspend fun updateMemo(memo: MemoType.Memo) {
-        localDataSource.updateMemo(memo.toMemoEntity())
-    }
-
-    override suspend fun checkIsCategory(categoryType: CategoryType, text: String): Boolean {
-        return localDataSource.checkIsCategory(categoryType.code,text) != null
-    }
-
-
-    override suspend fun deleteAllMemo() {
-        localDataSource.deleteAllMemo()
-    }
-
-    override suspend fun deleteMemo(id : Int) {
-        localDataSource.deleteMemo(id)
-    }
-
-    override suspend fun deleteCategory(text: String) {
-        localDataSource.deleteCategory(text)
     }
 }
