+++ core/domain/src/main/java/com/hegunhee/newsimplememoapp/domain/repository/CategoryRepository.kt
@@ -1,0 +1,14 @@
+package com.hegunhee.newsimplememoapp.domain.repository
+
+import com.hegunhee.newsimplememoapp.domain.model.CategoryType
+
+interface CategoryRepository {
+
+    suspend fun getAllCategoryByType(categoryType : CategoryType) : List<String>
+
+    suspend fun checkIsCategory(categoryType: CategoryType, text : String) : Boolean
+
+    suspend fun insertCategory(categoryType: CategoryType, text : String)
+
+    suspend fun deleteCategory(text : String)
+}
