+++ core/domain/src/main/java/com/hegunhee/newsimplememoapp/domain/usecase/category/CheckIsCategoryUseCase.kt
@@ -1,10 +1,10 @@
 package com.hegunhee.newsimplememoapp.domain.usecase.category
 
 import com.hegunhee.newsimplememoapp.domain.model.CategoryType
-import com.hegunhee.newsimplememoapp.domain.repository.MemoRepository
+import com.hegunhee.newsimplememoapp.domain.repository.CategoryRepository
 import javax.inject.Inject
 
-class CheckIsCategoryUseCase @Inject constructor(private val repository: MemoRepository) {
+class CheckIsCategoryUseCase @Inject constructor(private val repository: CategoryRepository) {
 
     suspend operator fun invoke(categoryType: CategoryType,text : String) :Boolean{
         return repository.checkIsCategory(categoryType,text)
