+++ core/domain/src/main/java/com/hegunhee/newsimplememoapp/domain/usecase/category/DeleteCategoryUseCase.kt
@@ -1,9 +1,9 @@
 package com.hegunhee.newsimplememoapp.domain.usecase.category
 
-import com.hegunhee.newsimplememoapp.domain.repository.MemoRepository
+import com.hegunhee.newsimplememoapp.domain.repository.CategoryRepository
 import javax.inject.Inject
 
-class DeleteCategoryUseCase @Inject constructor(private val repository: MemoRepository){
+class DeleteCategoryUseCase @Inject constructor(private val repository: CategoryRepository){
 
     suspend operator fun invoke(text : String) {
         repository.deleteCategory(text)
