+++ core/domain/src/main/java/com/hegunhee/newsimplememoapp/domain/usecase/category/GetAllCategoryByTypeUseCase.kt
@@ -1,10 +1,10 @@
 package com.hegunhee.newsimplememoapp.domain.usecase.category
 
 import com.hegunhee.newsimplememoapp.domain.model.CategoryType
-import com.hegunhee.newsimplememoapp.domain.repository.MemoRepository
+import com.hegunhee.newsimplememoapp.domain.repository.CategoryRepository
 import javax.inject.Inject
 
-class GetAllCategoryByTypeUseCase @Inject constructor(private val repository: MemoRepository) {
+class GetAllCategoryByTypeUseCase @Inject constructor(private val repository: CategoryRepository) {
 
     suspend operator fun invoke(categoryType: CategoryType) : List<String> {
         return repository.getAllCategoryByType(categoryType)
